Splitting the net into multiple work clusters
For each not yet visited place, create a new work cluster and add it to it. Add that place to the cluster and add its transitions. For each input place of these transitions, if it is not yet in the cluster, add it to the cluster. Repeat until the work cluster is fixed (no more places or transitions to add).